requirements.txt
# Web Frameworks
fastapi==0.104.1
uvicorn[standard]==0.24.0
flask==3.0.0
python-multipart==0.0.6  # For file uploads in FastAPI

# AI & OCR
openai==1.3.7
pdf2image==1.16.3
pillow==10.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9  # PostgreSQL driver

# Authentication & Security
pyjwt==2.8.0
bcrypt==4.1.1

# Validation
pydantic==2.5.2
pydantic[email]==2.5.2
email-validator==2.1.0

# Task Queue (for async processing)
celery==5.3.4
redis==5.0.1

# UI & Demos
gradio==4.7.1
pandas==2.1.3

# Utils
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
